requirements.txt
adal==1.2.2
aiohttp==3.6.2
ansible==2.9.6
apache-libcloud==2.8.0
applicationinsights==0.11.9
apturl==0.5.2
arandr==0.1.10
argcomplete==1.8.1
async-timeout==3.0.1
attrs==19.3.0
azure==4.0.0
azure-ai-textanalytics==1.0.0b1
azure-appconfiguration==1.0.0
azure-applicationinsights==0.1.0
azure-batch==8.0.0
azure-cli==2.0.81
azure-cli-core==2.0.81
azure-cli-telemetry==1.0.4
azure-cognitiveservices-anomalydetector==0.2.0
azure-cognitiveservices-formrecognizer==0.1.0
azure-cognitiveservices-knowledge-qnamaker==0.1.0
azure-cognitiveservices-language-luis==0.5.0
azure-cognitiveservices-language-spellcheck==2.0.0
azure-cognitiveservices-language-textanalytics==0.2.0
azure-cognitiveservices-personalizer==0.1.0
azure-cognitiveservices-search-autosuggest==0.2.0
azure-cognitiveservices-search-customimagesearch==0.2.0
azure-cognitiveservices-search-customsearch==0.3.0
azure-cognitiveservices-search-entitysearch==2.0.0
azure-cognitiveservices-search-imagesearch==2.0.0
azure-cognitiveservices-search-newssearch==2.0.0
azure-cognitiveservices-search-videosearch==2.0.0
azure-cognitiveservices-search-visualsearch==0.2.0
azure-cognitiveservices-search-websearch==2.0.0
azure-cognitiveservices-vision-computervision==0.5.0
azure-cognitiveservices-vision-contentmoderator==1.0.0
azure-cognitiveservices-vision-customvision==1.0.0
azure-cognitiveservices-vision-face==0.4.0
azure-common==1.1.24
azure-core==1.2.2
azure-cosmos==3.1.1
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.48
azure-devops==0.17.0
azure-eventgrid==1.3.0
azure-eventhub==5.0.0
azure-eventhub-checkpointstoreblob==1.0.0
azure-eventhub-checkpointstoreblob-aio==1.0.0
azure-functions-devops-build==0.0.22
azure-graphrbac==0.61.1
azure-identity==1.2.0
azure-keyvault==4.0.0
azure-keyvault-certificates==4.0.0
azure-keyvault-keys==4.1.0b1
azure-keyvault-secrets==4.1.0b1
azure-loganalytics==0.1.0
azure-mgmt==5.0.0a1
azure-mgmt-advisor==3.0.0
azure-mgmt-alertsmanagement==0.2.0rc2
azure-mgmt-apimanagement==0.1.0
azure-mgmt-appconfiguration==0.3.0
azure-mgmt-applicationinsights==0.2.0
azure-mgmt-appplatform==0.1.0
azure-mgmt-attestation==0.1.0
azure-mgmt-authorization==0.60.0
azure-mgmt-automation==0.1.1
azure-mgmt-azurestack==0.1.0
azure-mgmt-batch==7.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-botservice==0.2.0
azure-mgmt-cdn==4.1.0rc1
azure-mgmt-cognitiveservices==5.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==10.0.0
azure-mgmt-consumption==3.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc8
azure-mgmt-containerservice==8.2.0
azure-mgmt-cosmosdb==0.11.0
azure-mgmt-costmanagement==0.1.0
azure-mgmt-databoxedge==0.1.0
azure-mgmt-databricks==0.1.0
azure-mgmt-datafactory==0.8.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==4.0.0
azure-mgmt-datashare==0.1.0rc1
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devspaces==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==3.0.0
azure-mgmt-documentdb==0.1.3
azure-mgmt-edgegateway==0.1.0
azure-mgmt-eventgrid==3.0.0rc4
azure-mgmt-eventhub==3.0.0
azure-mgmt-frontdoor==0.3.0
azure-mgmt-hanaonazure==0.12.0
azure-mgmt-hdinsight==1.4.0
azure-mgmt-healthcareapis==0.1.0
azure-mgmt-hybridcompute==0.1.1
azure-mgmt-imagebuilder==0.3.0
azure-mgmt-iotcentral==2.0.0
azure-mgmt-iothub==0.10.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==2.0.0
azure-mgmt-kusto==0.5.0
azure-mgmt-labservices==0.1.1
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==4.0.0rc2
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-machinelearningservices==0.1.0
azure-mgmt-maintenance==0.1.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.2.1
azure-mgmt-media==2.1.0
azure-mgmt-mixedreality==0.1.0
azure-mgmt-monitor==0.7.0
azure-mgmt-msi==1.0.0
azure-mgmt-netapp==0.8.0
azure-mgmt-network==9.0.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-operationsmanagement==0.1.0
azure-mgmt-peering==0.1.0rc2
azure-mgmt-policyinsights==0.4.0
azure-mgmt-powerbidedicated==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-privatedns==0.1.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.4.0
azure-mgmt-recoveryservicesbackup==0.6.0
azure-mgmt-redis==7.0.0rc1
azure-mgmt-relay==0.2.0
azure-mgmt-reservations==0.7.0
azure-mgmt-resource==8.0.0
azure-mgmt-resourcegraph==2.0.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-security==0.3.0
azure-mgmt-serialconsole==0.1.0
azure-mgmt-servermanager==2.0.0
azure-mgmt-servicebus==0.6.0
azure-mgmt-servicefabric==0.4.0
azure-mgmt-signalr==0.3.0
azure-mgmt-sql==0.16.0
azure-mgmt-sqlvirtualmachine==0.5.0
azure-mgmt-storage==7.1.0
azure-mgmt-storagecache==0.2.0
azure-mgmt-storagesync==0.2.0
azure-mgmt-subscription==0.5.0
azure-mgmt-timeseriesinsights==0.1.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-vmwarecloudsimple==0.2.0
azure-mgmt-web==0.44.0
azure-multiapi-storage==0.2.4
azure-servicebus==0.50.2
azure-servicefabric==7.0.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
azure-storage-file==1.4.0
azure-storage-file-datalake==12.0.0b7
azure-storage-file-share==12.0.1
azure-storage-queue==1.4.0
azure-template==0.0.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fabric==2.5.0
fasteners==0.14.1
future==0.18.2
gpg===1.13.1-unknown
httplib2==0.14.0
humanfriendly==4.18
idna==2.8
invoke==1.3.0
isodate==0.6.0
javaproperties==0.6.0
Jinja2==2.10.1
jmespath==0.9.4
jsmin==2.2.2
jsondiff==1.1.1
keyring==18.0.1
knack==0.6.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mock==3.0.5
monotonic==1.5
msal==1.1.0
msal-extensions==0.1.3
msrest==0.6.10
msrestazure==0.6.2
multidict==4.7.3
netaddr==0.7.19
netifaces==0.10.4
ntlm-auth==1.1.0
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
pkginfo==1.4.2
ply==3.11
portalocker==1.5.1
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.8
PyAudio==0.2.11
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pysmb==1.2.8
PySmbClient==0.1.5
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pywinrm==0.3.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.0.0
requests-unixsocket==0.2.0
scp==0.13.0
SecretStorage==2.3.1
selinux==3.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
sshtunnel==0.1.4
systemd-python==234
tabulate==0.8.6
uamqp==1.2.6
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
vsts-cd-manager==1.0.2
wadllib==1.3.3
websocket-client==0.53.0
websockets==8.1
xkit==0.0.0
xmltodict==0.12.0
yarl==1.4.2
